A lot of small updates
After trial and error sessions we determined that Adamax was a better optimizer for the data. Adamax implemetation. Updated AddToMap method to take a data frame and list of rows to analyze instead of specifying certain column headers. UNSWAnalysis method changed to work with FULL datasets consisting of 4 csv files instead of pre-split training and test files. Included new models with Adamax optimizers.